+++ addons/it/sdi/identities.go
@@ -1,0 +1,10 @@
+package sdi
+
+import "github.com/invopop/gobl/cbc"
+
+const (
+	// Code managed by the CIPE (Interministerial Committee for Economic Planning) which characterises every public investment project (Individual Project Code)
+	IdentityTypeCUP cbc.Code = "CUP"
+	// Tender procedure identification code
+	IdentityTypeCIG cbc.Code = "CIG"
+)