+++ CHANGELOG.md
@@ -13,6 +13,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 ### Added
 
 - `cbc.Code`: new `Join` and `JoinWith` methods to help concatenate codes.
+- `it-sdi-v1`: added CIG and CUP identity type codes.
 
 ## [v0.201.0]